requirements.txt
boto3
flake8
fastapi
uvicorn
jinja2
aiofiles
werkzeug
psycopg2-binary
python-multipart
redis==3.5.3
celery==4.4.7
